Moved functionality from Location Show Page to Location Show Tile
