+++ app/assets/javascripts/application.js
@@ -12,7 +12,7 @@
 //
 //= require jquery
 //= require jquery_ujs
-//= //= require foundation
-require_tree .
+//= require foundation
+//= require_tree .
 
 $(function(){ $(document).foundation(); });
