+++ config/routes.rb
@@ -8,5 +8,5 @@ Rails.application.routes.draw do
     end
   end
 
-  get '/locations/(:id)', to: 'homes#index'
+  get '/locations/:id', to: 'homes#index'
 end
